part 2
changed the parser to store the data as a dictionary keyed by the names.
Then calculated the total weights on each as it filters down the tree,
and checks the balance of each as it goes. For the first unbalanced disk
it says what the weight needed to balance it should be.